+++ src/mesh/Mesh.cpp
@@ -22,7 +22,8 @@ Mesh::Mesh(
     : _name(name),
       _dimensions(dimensions),
       _flipNormals(flipNormals),
-      _id(id)
+      _id(id),
+      _boundingBox(dimensions)
 {
   PRECICE_ASSERT((_dimensions == 2) || (_dimensions == 3), _dimensions);
   PRECICE_ASSERT(_name != std::string(""));
@@ -197,10 +198,18 @@ bool Mesh::isValidEdgeID(int edgeID) const
 void Mesh::allocateDataValues()
 {
   PRECICE_TRACE(_vertices.size());
+  const auto expectedCount = _vertices.size();
+  using SizeType = decltype(expectedCount);
   for (PtrData data : _data) {
-    int total          = _vertices.size() * data->getDimensions();
-    int leftToAllocate = total - data->values().size();
-    if (leftToAllocate > 0) {
+    const SizeType expectedSize = expectedCount * data->getDimensions();
+    const auto actualSize = static_cast<SizeType>(data->values().size());
+    // Shrink Buffer
+    if (expectedSize < actualSize) {
+        data->values().resize(expectedSize);
+    }
+    // Enlarge Buffer
+    if (expectedSize > actualSize) {
+      const auto leftToAllocate = expectedSize - actualSize;
       utils::append(data->values(), (Eigen::VectorXd) Eigen::VectorXd::Zero(leftToAllocate));
     }
     PRECICE_DEBUG("Data " << data->getName() << " now has " << data->values().size() << " values");
@@ -210,19 +219,13 @@ void Mesh::allocateDataValues()
 void Mesh::computeBoundingBox()
 {
   PRECICE_TRACE(_name);
-  BoundingBox boundingBox(_dimensions,
-                          std::make_pair(std::numeric_limits<double>::max(),
-                                         std::numeric_limits<double>::lowest()));
+  BoundingBox bb(_dimensions);
   for (const Vertex &vertex : _vertices) {
-    for (int d = 0; d < _dimensions; d++) {
-      boundingBox[d].first  = std::min(vertex.getCoords()[d], boundingBox[d].first);
-      boundingBox[d].second = std::max(vertex.getCoords()[d], boundingBox[d].second);
-    }
+    bb.expandBy(vertex);
   }
-  for (int d = 0; d < _dimensions; d++) {
-    PRECICE_DEBUG("BoundingBox, dim: " << d << ", first: " << boundingBox[d].first << ", second: " << boundingBox[d].second);
-  }
-  _boundingBox = std::move(boundingBox);
+  _boundingBox = std::move(bb);
+  PRECICE_DEBUG("Bounding Box, " << _boundingBox);
+  
 }
 
 void Mesh::computeState()
@@ -404,18 +407,9 @@ void Mesh::addMesh(
   meshChanged(*this);
 }
 
-const Mesh::BoundingBox Mesh::getBoundingBox() const
+const BoundingBox &Mesh::getBoundingBox() const
 {
   return _boundingBox;
-}
-
-const std::vector<double> Mesh::getCOG() const
-{
-  std::vector<double> cog(_dimensions);
-  for (int d = 0; d < _dimensions; d++) {
-    cog[d] = (_boundingBox[d].second - _boundingBox[d].first) / 2.0 + _boundingBox[d].first;
-  }
-  return cog;
 }
 
 bool Mesh::operator==(const Mesh &other) const