+++ src/precice/config/ParticipantConfiguration.cpp
@@ -156,10 +156,12 @@ ParticipantConfiguration::ParticipantConfiguration(
                             "bind it automatically.");
     tagMaster.addAttribute(attrPort);
 
-    auto attrNetwork = makeXMLAttribute(ATTR_NETWORK, "lo")
+    auto attrNetwork = makeXMLAttribute(ATTR_NETWORK, utils::networking::loopbackInterfaceName())
                            .setDocumentation(
-                               "Network name to be used for socket communiation. "
-                               "Default is \"lo\", i.e., the local host loopback.");
+                               "Interface name to be used for socket communiation. "
+                               "Default is the cannonical name of the loopback interface of your platform. "
+                               "Might be different on supercomputing systems, e.g. \"ib0\" "
+                               "for the InfiniBand on SuperMUC. ");
     tagMaster.addAttribute(attrNetwork);
 
     auto attrExchangeDirectory = makeXMLAttribute(ATTR_EXCHANGE_DIRECTORY, "")
@@ -232,8 +234,11 @@ void ParticipantConfiguration::xmlTagCallback(
       stream << "Safety Factor must be positive or 0";
       throw std::runtime_error{stream.str()};
     }
-    bool          provide = tag.getBooleanAttributeValue(ATTR_PROVIDE);
-    mesh::PtrMesh mesh    = _meshConfig->getMesh(name);
+    bool provide = tag.getBooleanAttributeValue(ATTR_PROVIDE);
+    if (_participants.back()->getName() == from) {
+      PRECICE_CHECK(provide, "Participant \"" << context.name << "\" cannot use mesh \"" << name << "\" from itself. Use the \"from\"-field to specify which participant has to communicate the mesh to \"" << context.name << "\".");
+    }
+    mesh::PtrMesh mesh = _meshConfig->getMesh(name);
     if (mesh.get() == nullptr) {
       std::ostringstream stream;
       stream << "Participant \"" << _participants.back()->getName()