+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file. For future 
 
 ## 2.1.0
 
+- Refactor `cplscheme::BaseCouplingScheme` and derived classes. Introduce `cplscheme::BiCouplingScheme`.
+- Refactor `mesh::BoundingBox` into seperate class
 - Removed deprecated and untested Manifold Mapping. API functions `hasToEvaluateSurrogateModel` and `hasToEvaluateFineModel` remain as nop stubs.
 - Change the CMake FindNumPy module to only consider information based on the selected python interpreter.
 - Implemented a new action to sum up data values, `action:summation`.